refactor: commons folder remove unused & dependent
