+++ webclient/app/src/app/App.jsx
@@ -4,6 +4,7 @@ import MainContentRouter from "./pages/MainContentRouter";
 import {CssBaseline} from "@mui/material";
 import AppStyles from "./AppStyles";
 import ErrorHandler from "./commons/errorHandler/ErrorHandler";
+import {appMenuItems} from "./AppConfig";
 
 function App() {
     const appStyles = AppStyles();
@@ -13,7 +14,7 @@ function App() {
             <ErrorHandler>
                 <div className={appStyles.background}>
                     <CssBaseline/>
-                    <AppHeader/>
+                    <AppHeader menuItems={appMenuItems}/>
                     <MainContentRouter/>
                 </div>
             </ErrorHandler>
