+++ webclient/app/src/app/AppConfig.js
@@ -1,0 +1,6 @@
+const appMenuItems = [
+    {title: "apps.title", link: "/"},
+    {title: "apptemplates.title", link: "/apptemplates/all"}
+];
+
+export {appMenuItems};
