+++ webclient/app/src/app/commons/appHeader/AppHeader.jsx
@@ -7,7 +7,8 @@ import {useHistory} from "react-router-dom";
 import {useTranslation} from "react-i18next";
 import {Logout} from "@mui/icons-material";
 
-function AppHeader() {
+function AppHeader(props) {
+    const {menuItems} = props;
     const appHeaderStyles = AppHeaderStyles();
     const history = useHistory();
     const {t} = useTranslation();
@@ -18,10 +19,10 @@ function AppHeader() {
                 <Toolbar className={appHeaderStyles.toolbar}>
                     <img className={appHeaderStyles.menuLogoImg} src={logo} alt="Logo of lirejarp"/>
                     <div className={appHeaderStyles.spacer}/>
-                    <Button color="secondary" disableRipple className={appHeaderStyles.linkButton}
-                        onClick={() => history.push("/")}>{t("apps.title")}</Button>
-                    <Button color="secondary" disableRipple className={appHeaderStyles.linkButton}
-                        onClick={() => history.push("/apptemplates/all")}>{t("apptemplates.title")}</Button>
+                    {menuItems.map(item => (
+                        <Button key={item.title} color="secondary" disableRipple className={appHeaderStyles.linkButton}
+                            onClick={() => history.push(item.link)}>{t(item.title)}</Button>
+                    ))}
                     <IconButton color="secondary" disableRipple className={appHeaderStyles.linkButton}
                         onClick={() => history.push("/logout")}><Logout/></IconButton>
                 </Toolbar>
