+++ webclient/app/src/app/commons/inputfields/multipleSelectChip/MultipleSelectChipStyles.js
