+++ webclient/app/src/app/features/appTemplates/appTemplateCard/AppTemplateCard.jsx
@@ -27,7 +27,7 @@ import AppTemplateDialog from "./../appTemplateDialog/AppTemplateDialog";
 import AppTemplateRest from "../../../services/AppTemplateRest";
 import GitRest from "../../../services/GitRest";
 import ConfirmationDialog from "../../../commons/confirmationDialog/ConfirmationDialog";
-import GitDataButton from "../../../commons/gitDownloadButton/GitDataButton";
+import GitDataButton from "../../../commons/buttons/gitDownloadButton/GitDataButton";
 import {useSnackbar} from "notistack";
 
 const ExpandMore = styled(props => {
