+++ webclient/app/src/app/features/apps/entities/relationshipAccordion/RelationshipAccordion.jsx
@@ -17,7 +17,7 @@ import PropTypes from "prop-types";
 import RelationshipStyles from "./RelationshipStyles";
 import {useTranslation} from "react-i18next";
 import {RelationshipType} from "../Relationship";
-import ValidatedTextField from "../../../../commons/validatedTextField/ValidatedTextField";
+import ValidatedTextField from "../../../../commons/inputfields/validatedTextField/ValidatedTextField";
 import RegexConfig from "../../../../../regexConfig";
 
 function RelationshipAccordion(props) {
